requirements.txt
amqp==5.3.1
asgiref==3.8.1
attrs==25.3.0
autobahn==24.4.2
Automat==25.4.16
billiard==4.2.1
celery==5.5.3
cffi==1.17.1
click==8.2.1
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
constantly==23.10.4
cron-descriptor==1.4.5
cryptography==45.0.6
daphne==4.2.1
dj-database-url==3.0.1
Django==5.2.1
django-celery-beat==2.8.1
django-timezone-field==7.1
django_celery_results==2.6.0
flower==2.0.1
honcho==2.0.0
humanize==4.12.3
hyperlink==21.0.0
idna==3.10
incremental==24.7.2
kombu==5.5.4
packaging==25.0
prometheus_client==0.22.1
prompt_toolkit==3.0.51
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.22
pyOpenSSL==25.1.0
python-crontab==3.2.0
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
pytz==2025.2
redis==6.2.0
service-identity==24.2.0
six==1.17.0
sqlparse==0.5.3
tornado==6.5.1
Twisted==25.5.0
txaio==25.6.1
typing_extensions==4.15.0
tzdata==2025.2
vine==5.1.0
wcwidth==0.2.13
zope.interface==7.2
